Skip `cd %BUILD_DIR%\src` on cygwin.
This is the very very last command that executes in the Cygwin build, since the next line is disabled. Of course, changing a directory is irrelevant. But that directory doesn't exist after the Cygwin build. So it fails. So the build failed ON THE LAST LINE THAT DIDN'T MATTER ANYWAY.